+++ metaconfigure/description.json
@@ -469,6 +469,9 @@
     "src/njoy21/input/HEATR/Card2.hpp",
     "src/njoy21/input/HEATR/Card5.hpp",
     "src/njoy21/input/HEATR/Card4/Mta.hpp",
+    "src/njoy21/input/HEATR/src/buildCard3.hpp",
+    "src/njoy21/input/HEATR/src/buildControlTuple.hpp",
+    "src/njoy21/input/HEATR/src/ctor.hpp",
     "src/njoy21/input/HEATR/Card3/Mtk.hpp",
     "src/njoy21/input/HEATR/Card2/Npk.hpp",
     "src/njoy21/input/HEATR/Card2/Ntemp.hpp",
